% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_evaluation.R
\name{AIC_gamma_HMM}
\alias{AIC_gamma_HMM}
\title{Calculate AIC for fitted gamma HMM}
\usage{
AIC_gamma_HMM(mllk, N, p)
}
\arguments{
\item{mllk}{minimum negative Log-Likelihood, output of \code{fit_gamma_hmm}.}

\item{N}{Number of states of the HMM.}

\item{p}{Degree of autocorrelation of the HMM (0 - no autocorrelation).}
}
\value{
AIC of fitted model.
}
\description{
Calculate the AIC criterion for a fitted gamma HMM 
(with or without autocorrelation).
}
