% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_decoding.R
\name{allprobs}
\alias{allprobs}
\title{Matrix of all probabilities}
\usage{
allprobs(x, dists, autocor = 0, params, N, p)
}
\arguments{
\item{x}{Data matrix the model was fitted to.}

\item{dists}{Vector of distributions in R-jargon.}

\item{autocor}{list of autocorrelation vectors, respecting order of dists.}

\item{params}{List of optimized parameters, returned by fitting the HMM.}

\item{N}{Number of states.}

\item{p}{Vector of degree of autocorrelation for each distribution (0 = no autocorrelation).}
}
\value{
Matrix of all probabilities.
}
\description{
Calculate matrix of all probabilities for an AR(p) HMM with or without autocorrelation.
}
