% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_evaluation.R
\name{BIC_gamma_HMM}
\alias{BIC_gamma_HMM}
\title{Calculate BIC for fitted gamma HMM}
\usage{
BIC_gamma_HMM(mllk, N, p, data)
}
\arguments{
\item{mllk}{minimum negative Log-Likelihood, output of \code{fit_gamma_hmm}.}

\item{N}{Number of states of the HMM.}

\item{p}{Degree of autocorrelation of the HMM (0 - no autocorrelation).}

\item{data}{Data vector the data was fitted to, to determine \eqn{\log(n)}.}
}
\value{
BIC of fitted model.
}
\description{
Calculate the BIC criterion for a fitted gamma HMM 
(with or without autocorrelation).
}
