% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{boxplot_sim_params}
\alias{boxplot_sim_params}
\title{Full simulation boxplots of estimated parameters}
\usage{
boxplot_sim_params(
  gamma_mu,
  gamma_sigma,
  vm_mu,
  vm_kappa,
  gamma_auto = 0,
  vm_auto = 0,
  true_values,
  nstates = 2,
  vm_head_pos = 0
)
}
\arguments{
\item{gamma_mu}{Fitted values of the parameter mu of the gamma distribution.}

\item{gamma_sigma}{Fitted values of the parameter sigma of the gamma distribution.}

\item{vm_mu}{Fitted values of the parameter mu of the von Mises distribution.}

\item{vm_kappa}{Fitted values of the parameter kappa of the von Mises distribution.}

\item{gamma_auto}{Fitted values of the autoregression parameters of the gamma distribution. 
0 if no autoregression was fitted.}

\item{vm_auto}{Fitted values of the autoregression parameters of the von Mises distribution.
0 if no autoregression was fitted.}

\item{true_values}{True parameter values, order: [gamma_mu, gamma_sigma, vm_mu, vm_kappa, gamma_auto, vm_auto]}

\item{nstates}{Number of states fitted (default: 2). Works only for 2 states!}

\item{vm_head_pos}{optional, determines position of heading of von Mises distribution. 
Only gets used for fitted models without autoregression}
}
\description{
Generate a boxplot of all fitted parameters, generated by fitting an AR(p)-HMM
to data. Includes a comparison to the true value. Currently only works
for a simulation containing one gamma distributed and one von Mises distributed
variable.
}
