% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{circ_vm_viz}
\alias{circ_vm_viz}
\title{Circular KDE and density of von Mises distribution}
\usage{
circ_vm_viz(mu, kappa, delta = 1, data = NULL, sum_dist = FALSE, leg = FALSE)
}
\arguments{
\item{mu}{Vector of parameter mu for von Mises distribution.}

\item{kappa}{Vector of parameter mu for von Mises distribution.}

\item{delta}{Optional, weights of the distributions (equal to delta in HMMs).}

\item{data}{Optional, provide data vector, KDE of this data will be plotted.}

\item{sum_dist}{Optional, sum weighted distributions.}

\item{leg}{Optional, include legend.}
}
\description{
Generate a circular visualization of the density of one or several von Mises
distributed variables. Also, if data is supplied, generate the kernel density
estimator of the data for comparison.
}
