% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dens_vm}
\alias{dens_vm}
\title{Density of von Mises distribution}
\usage{
dens_vm(x, theta, autocor_ind = 0, autocor = 0, p = 0)
}
\arguments{
\item{x}{Data.}

\item{theta}{Named list of parameters of von Mises distribution.}

\item{autocor_ind}{Matrix, indices to consider when computing the autoregressive parameters.}

\item{autocor}{Vector, j-th row of autocor matrix, with one value for each time lag considered.}

\item{p}{Int, degree of autoregression}
}
\value{
Vector of densities of von Mises distribution.
}
\description{
Function to calculate the density of different values for a von Mises distributed random variable
(with possible autoregression in \eqn{mu}).
}
