% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_fit.R
\name{fit_arp_model}
\alias{fit_arp_model}
\title{Fit an AR(p) HMM to data}
\usage{
fit_arp_model(mllk, data, theta.star, N, p, dists)
}
\arguments{
\item{mllk}{Negative log-Likelihood function that should be minimized.}

\item{data}{Data that should be fitted using the HMM.}

\item{theta.star}{Unconstrained parameter vector (has to be provided in suitable form).
Attention: -Inf values (resulting e.g. from supplying autocorrelation = 0) are not possible
for the optimization function.}

\item{N}{Number of states.}

\item{p}{Vector of degree of autocorrelation for each distribution, 0=no autocorrelation.}

\item{dists}{Vector containing abbreviated names (in R-jargon) of the distributions 
to be considered in the Likelihood computation.}
}
\value{
List, containing minimal value of negative log-Likelihood, Gamma, 
        delta, (autocorrelation, depending on degree), mu, sigma.
}
\description{
Fit an AR(p) HMM with autocorrelation to data 
using a specified function to compute the negative 
Log-Likelihood. This function gets minimized using the function \code{optim}.
It returns the estimated parameters of the fitted model. The distribution of the HMM
can be specified using the parameter \code{dist}.
}
