% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_fit.R
\name{fit_arp_model_gamma_mu}
\alias{fit_arp_model_gamma_mu}
\title{Fit an AR(p)-gamma HMM to data}
\usage{
fit_arp_model_gamma_mu(mllk, data, theta.star, N, p)
}
\arguments{
\item{mllk}{Negative log-Likelihood function that should be minimized.}

\item{data}{Data that should be fitted using the HMM.}

\item{theta.star}{Unconstrained parameter vector (has to be provided in suitable form).}

\item{N}{Number of states.}

\item{p}{Degree of autocorrelation, 0 is possible for no autocorrelation.}
}
\value{
List, containing minimal value of negative log-Likelihood, Gamma, 
        delta, (autocorrelation, depending on degree), mu, sigma.
}
\description{
Fit an AR(p)-gamma HMM with autocorrelation in the parameter \eqn{\mu} to data 
using a specified function to compute the negative 
Log-Likelihood. This function gets minimized using the function \code{optim}.
It returns the estimated parameters of the fitted model.
}
