% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_mllk.R
\name{mllk}
\alias{mllk}
\title{Compute negative Log-Likelihood}
\usage{
mllk(theta.star, dists, x, N, p)
}
\arguments{
\item{theta.star}{Named list of parameters, containing: 1) Full TPM, 2) List of 
autocorrelation parameters, one vector entry for each distribution, 
3) List with sublist for each distribution, containing named 
parameters (N entries for each parameter of each distribution)}

\item{dists}{Vector containing abbreviated names (in R-jargon) of the distributions 
to be considered in the Likelihood computation.}

\item{x}{Data vector or matrix for which the negative Log-Likelihood should be computed.}

\item{N}{Number of states.}

\item{p}{Vector of degree of autocorrelation for each distribution, 0=no autocorrelation.}
}
\value{
Negative Log-Likelihood.
}
\description{
Compute the negative Log-Likelihood, using one or several specified distributions.
The distributions have to be specified by their commonly known abbreviation in R, 
e.g. one of ['gamma', 'vm', 'pois', 'binom',...].
The named list of parameters (one value for each parameter and for each state) have to be
in suitable form, i.e. they have to respect the natural parameter boundaries of the distributions.
In the likelihood computation, contemporaneus independence is assumed.
}
