% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_mllk.R
\name{mllk_gamma_hmm}
\alias{mllk_gamma_hmm}
\title{Calculate negative Log-Likelihood of gamma HMM}
\usage{
mllk_gamma_hmm(theta.star, x, N, p = 0)
}
\arguments{
\item{theta.star}{Unconstrained parameter vector (has to be provided in suitable form).}

\item{x}{Data vector for which the negative Log-Likelihood should be computed.}

\item{N}{Number of states.}

\item{p}{Degree of autocorrelation, here always = 0.}
}
\value{
Minus Log-Likelihood
}
\description{
Calculate the negative Log-Likelihood of a \emph{normal} gamma HMM 
(without autocorrelation). The forward method using standardized forward 
variables is used here.
}
