% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_mllk.R
\name{mllk_vonMises_arp}
\alias{mllk_vonMises_arp}
\title{Calculate negative Log-Likelihood of AR(p)-von Mises HMM}
\usage{
mllk_vonMises_arp(theta.star, x, N, p)
}
\arguments{
\item{theta.star}{Unconstrained parameter vector (has to be provided in suitable form:
Gamma, mu, kappa, autocorrelation)!}

\item{x}{Data vector for which the negative Log-Likelihood should be computed.}

\item{N}{Number of states.}

\item{p}{Degree of autocorrelation, 0=no autocorrelation}
}
\value{
Minus Log-Likelihood
}
\description{
Calculate the negative Log-Likelihood of an von Mises HMM with \emph{AR(p)} in the
parameter \eqn{\mu} (and \eqn{\kappa} in the future?).
The forward method using standardized forward variables is used here.
}
