% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{plot_data}
\alias{plot_data}
\title{Plot data}
\usage{
plot_data(data, name = "none", title = "none")

plot_data(data, name = "none", title = "none")
}
\arguments{
\item{data}{Data vector to be plotted}

\item{name}{optional, y-axis label, if "none" than "data"}

\item{title}{optional, title of the plot, if "none", then no title}
}
\description{
Plot a data vector

Plot a data vector
}
