% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{plot_decoded_data}
\alias{plot_decoded_data}
\title{Plot decoded data}
\usage{
plot_decoded_data(
  data,
  states,
  col,
  name = "none",
  title = "none",
  legend = TRUE
)
}
\arguments{
\item{data}{Data vector to be plotted.}

\item{states}{Vector of ecoded states.}

\item{col}{Color of the states.}

\item{name}{optional, y-axis label, if "none" than "data".}

\item{title}{optional, title of the plot, if "none", then no title.}

\item{legend}{optional, bool, specifies if there should be a legend.}
}
\description{
Plot a data vector, colored with regard to the Viterbi decoded states.
}
