% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{plot_fitted_dist}
\alias{plot_fitted_dist}
\title{Density plot of estimated distributions}
\usage{
plot_fitted_dist(
  data,
  dist,
  param,
  N,
  delta,
  title = "none",
  breaks = 30,
  xlab = "x",
  legend = TRUE,
  xlim = "none"
)
}
\arguments{
\item{data}{Input data for histogram.}

\item{dist}{string, indicates the kind of distribution, one of ['gamma', 'von Mises'].}

\item{param}{Vector of distribution parameters, in customary form, e.g. for 2 state gamma
HMM: c(\eqn{\mu1,\mu2,\sigma1,\sigma2}).}

\item{N}{int, number of states (or weighted distributions) to be plotted.}

\item{delta}{Input vector for parameter delta.}

\item{title}{optional, title of the plot, if "none", then no title.}

\item{breaks}{optional, specify number of breaks of the histogram.}

\item{xlab}{optional, label of x-axis.}

\item{legend}{optional, plot legend (default: TRUE).}

\item{xlim}{optional, set a limit for x-axis for better visibility of the data.}
}
\description{
Plots the estimted distributions along with a histogram of the values.
}
