% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_sim_functions.R
\name{plot_fitted_gamma_dist}
\alias{plot_fitted_gamma_dist}
\title{Density plot of estimated gamma distributions}
\usage{
plot_fitted_gamma_dist(data, mu, sigma, delta, title = "none")
}
\arguments{
\item{data}{Input data for histogram}

\item{mu}{Input vector for parameter mu}

\item{sigma}{Input vector for parameter sigma}

\item{delta}{Input vector for parameter delta}

\item{title}{optional, title of the plot, if "none", then no title}
}
\description{
Plots the estimted gamma distributions along with a histogram of the values.
}
