% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{plot_states}
\alias{plot_states}
\title{Plot states decoded by Viterbi algorithm}
\usage{
plot_states(states, names = "none", title = TRUE)
}
\arguments{
\item{states}{Vector of states (natural numbers, range 1-N).}

\item{names}{optional, specifies the names of the states. If "none", state 1-N will be used.}

\item{title}{bool, indicator if the plot should habe a title.}
}
\description{
This function plots the states decoded by the Viterbi algorithm.
}
