% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_pseudores.R
\name{pseudores_arp}
\alias{pseudores_arp}
\title{Compute pseudo residuals}
\usage{
pseudores_arp(mod, data, N, p)
}
\arguments{
\item{mod}{Fitted AR(p)-HMM object.}

\item{data}{Data the model is fitted on.}

\item{N}{Number of states.}

\item{p}{Vector of degree of autoregression for each distribution (0 = no autoregression).}
}
\value{
Pseudo residuals for step length and turning angles.
}
\description{
Compute the pseudo residuals of an AR(p)-HMM.
Adapted from moveHMM.
Attention: Only works for 2-dim with 1st variable gamma distributed step lengths 
and 2nd variable von Mises distributed turning angles!
}
