% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dists.R
\name{sample_gamma}
\alias{sample_gamma}
\title{Sample from gamma distribution}
\usage{
sample_gamma(n, mu, sigma, cv)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Parameter \eqn{\mu}.}

\item{sigma}{Parameter \eqn{\sigma}.}
}
\value{
Sample(s)
}
\description{
Function to sample from gamma distributed random variable.
}
