% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_simulate.R
\name{sample_hmm_ar1}
\alias{sample_hmm_ar1}
\title{Simulate data from a Gamma HMM with AR(1) structure}
\usage{
sample_hmm_ar1(n_samples, delta, Gamma, N, mu, sigma, autocor)
}
\arguments{
\item{n_samples}{Number of samples to generate.}

\item{delta}{Initial distribution of the Markov chain.}

\item{Gamma}{Transition probability matrix of the Markov chain.}

\item{N}{Number of states.}

\item{mu}{Parameter vector for mu of the gamma distribution.}

\item{sigma}{Parameter vector for sigma of the gamma distribution.}

\item{autocor}{Parameter vector for the autocorrelation coefficient.}
}
\value{
Tupel of states and data of the HMM.
}
\description{
Simulate data from a Gamma HMM with AR(1) structure in parameters \eqn{\mu} and \eqn{\sigma}.
}
