% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_simulate.R
\name{sample_hmm_arp_mu}
\alias{sample_hmm_arp_mu}
\title{Simulate data from a Gamma HMM with AR(p) structure}
\usage{
sample_hmm_arp_mu(n_samples, delta, Gamma, N, mu, sigma, autocor, p)
}
\arguments{
\item{n_samples}{Number of samples to generate.}

\item{delta}{Initial distribution of the Markov chain.}

\item{Gamma}{Transition probability matrix of the Markov chain.}

\item{N}{Number of states.}

\item{mu}{Parameter vector for mu of the gamma distribution.}

\item{sigma}{Parameter vector for sigma of the gamma distribution.}

\item{autocor}{Parameter vector for the autocorrelation coefficient. 
Has to match p, in the order \eqn{\phi_{t-p},\dots,\phi_{t-1}}
where \eqn{\phi} is the vector of autocorrelation coefficients
for one specific time lag (one value for each state).}

\item{p}{Degree of autocorrelation.}
}
\value{
Tupel of states and data of the HMM.
}
\description{
Simulate data from a Gamma HMM with AR(p) structure in the parameter \eqn{\mu}.
}
