% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_sim_functions.R
\name{sample_hmm_normal}
\alias{sample_hmm_normal}
\title{Simulate data from a Gamma HMM}
\usage{
sample_hmm_normal(n_samples, delta, Gamma, N, mu, sigma)
}
\arguments{
\item{n_samples}{Number of samples to generate.}

\item{delta}{Initial distribution of the Markov chain.}

\item{Gamma}{Transition probability matrix of the Markov chain.}

\item{N}{Number of states.}

\item{mu}{Parameter vector for mu of the gamma distribution.}

\item{sigma}{Parameter vector for sigma of the gamma distribution.}
}
\value{
Tupel of states and data of the HMM.
}
\description{
Simulate data from a Gamma HMM
}
