% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dists.R
\name{sample_norm}
\alias{sample_norm}
\title{Sample from normal distribution}
\usage{
sample_norm(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples}

\item{mu}{Parameter \eqn{\mu}.}

\item{sigma}{Parameter \eqn{\sigma}.}
}
\value{
Sample(s)
}
\description{
Function to sample from normal distributed random variable
}
