% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dists.R
\name{sample_vm}
\alias{sample_vm}
\title{Sample from von Mises distribution}
\usage{
sample_vm(n, mu, kappa)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Parameter \eqn{\mu}.}

\item{kappa}{Parameter \eqn{\kappa}.}
}
\value{
Sample(s)
}
\description{
Function to sample from von Mises distributed random variable.
}
