% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_simulate.R
\name{simulate_track}
\alias{simulate_track}
\title{Generate track from simulated data}
\usage{
simulate_track(simulated_data, coord_start = c(0, 0), direction_start = 0)
}
\arguments{
\item{simulated_data}{DataFrame or matrix, simulated step lengths and turning angles.}

\item{coord_start}{Vector, initial coordinates to start in, default: (0,0).}

\item{direction_start}{int, initial direction, the first turning angle should be relative to. 
Scale: [-pi,pi), default: 0}
}
\value{
Data frame of coordinates, step length and turning angles, similar to 
         output from prepData in moveHMM.
}
\description{
Generate movement path coordinates from simulated data (step length and turning
angle) of an AR(p)-HMM.
}
