% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_params.R
\name{starize}
\alias{starize}
\title{Natural parameters to working parameters (theta->theta.star)}
\usage{
starize(theta, N, p, dists)
}
\arguments{
\item{theta}{Vector of natural parameters.}

\item{N}{Number of states of the HMM.}

\item{p}{Vector of degree of autocorrelation within the distributions.}

\item{dists}{Vector of the distributions in the HMM.}
}
\value{
Vector of working parameters of the HMM.
}
\description{
Make the natural parameters of HMMs to their working parameters for optimization (theta -> theta.star).
Currently only works for distributions with exactly two parameters.
}
