% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_params.R
\name{starting_params_opt}
\alias{starting_params_opt}
\title{Generate starting parameters for a distribution}
\usage{
starting_params_opt(data, dist, N)
}
\arguments{
\item{data}{Vector of natural parameters.}

\item{dist}{Distribution in R-jargon.}

\item{N}{Number of states of the HMM.}
}
\value{
Vector of starting parameters.
}
\description{
Generate meaningful starting parameters for the optimization of a specified distribution using 
common sense and the supplied data.
}
