% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_decoding.R
\name{viterbi_arp}
\alias{viterbi_arp}
\title{Global decoding for AR(p) HMM using Viterbi}
\usage{
viterbi_arp(x, Gamma, delta, dists, autocor = 0, params, N, p)
}
\arguments{
\item{x}{Data matrix the model was fitted to.}

\item{Gamma}{Transition probability matrix (full matrix, not only off diagonal entries).}

\item{delta}{Stationary distribution.}

\item{dists}{Vector of distributions in R-jargon.}

\item{autocor}{list of autocorrelation vectors, respecting order of dists.}

\item{params}{List of optimized parameters, returned by fitting the HMM.}

\item{N}{Number of states.}

\item{p}{Vector of degree of autocorrelation for each distribution (0 = no autocorrelation).}
}
\value{
Estimated states using Viterbi.
}
\description{
Global decoding for an AR(p) HMM with or without autocorrelation 
using the Viterbi algorithm.
}
