% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_decoding.R
\name{viterbi_gamma_arp}
\alias{viterbi_gamma_arp}
\title{Global decoding for AR(p)-gamma HMM using Viterbi}
\usage{
viterbi_gamma_arp(x, Gamma, delta, autocor = 0, mu, sigma, N, p)
}
\arguments{
\item{x}{Data vector the model was fitted to.}

\item{Gamma}{Transition probability matrix (full matrix, not only off diagonal entries).}

\item{delta}{Stationary distribution.}

\item{autocor}{default 0, Autocorrelation vector, in suitable form.}

\item{mu}{Optimized vector of the mu parameter in the gamma distribution.}

\item{sigma}{Optimized vector of the mu parameter in the gamma distribution.}

\item{N}{Number of states.}

\item{p}{Degree of autocorrelation (0 - no autocorrelation).}
}
\value{
Estimated states using Viterbi.
}
\description{
Global decoding for an AR(p)-gamma HMM with autocorrelation in the parameters \eqn{\mu}
and \eqn{\sigma}
using the Viterbi algorithm.
}
