% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dens_gamma}
\alias{dens_gamma}
\title{Density of gamma distribution}
\usage{
dens_gamma(x, theta, autocor_ind = 0, autocor = 0, p = 0)
}
\arguments{
\item{x}{Data.}

\item{theta}{Named list of parameters of gamma distribution.}

\item{autocor_ind}{Matrix, indices of to consider when computing the autocorrelated parameters.}

\item{p}{Int, degree of autocorrelation.}
}
\value{
Densities of gamma distribution.
}
\description{
Function to calculate the density of different values for a gamma distributed random variable 
(with possible autocorrelation in \eqn{mu} and \eqn{sigma}, due to the constant coefficient of variation).
}
