% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dens_norm}
\alias{dens_norm}
\title{Density of normal distribution}
\usage{
dens_norm(x, theta, autocor_ind, autocor, p)
}
\arguments{
\item{x}{Data}

\item{theta}{Named list of parameters of normal distribution.}

\item{autocor_ind}{Matrix, indices of to consider when computing the autocorrelated parameters.}

\item{p}{Int, degree of autocorrelation.}
}
\value{
Densities of normal distribution.
}
\description{
Function to calculate the density of different values for a normal distributed random variable
(currently only with possible autocorrelation in parameter \eqn{mu}).
}
