% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{dens_vm}
\alias{dens_vm}
\title{Density of von Mises distribution}
\usage{
dens_vm(x, theta, autocor_ind = 0, autocor = 0, p = 0)
}
\arguments{
\item{x}{Data}

\item{theta}{Named list of parameters of von Mises distribution.}

\item{autocor_ind}{Matrix, indices of to consider when computing the autocorrelated parameters.}

\item{autocor}{Vector, j-th column of autocor matrix, with one value for each time lag considered}

\item{p}{Int, degree of autocorrelation.}
}
\value{
Densities of von Mises distribution.
}
\description{
Function to calculate the density of different values for a von Mises distributed random variable.
}
