% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_mllk.R
\name{mllk_arp}
\alias{mllk_arp}
\title{Calculate negative Log-Likelihood of AR(p)-gamma HMM}
\usage{
mllk_arp(theta.star, x, N, p)
}
\arguments{
\item{theta.star}{Unconstrained parameter vector (has to be provided in suitable form).}

\item{x}{Data vector for which the negative Log-Likelihood should be computed.}

\item{N}{Number of states.}

\item{p}{Degree of autocorrelation.}
}
\value{
Minus Log-Likelihood
}
\description{
Calculate the negative Log-Likelihood of an gamma HMM with \emph{AR(p)} in the
parameters \eqn{\mu} and \eqn{\sigma}.
The forward method using standardized forward variables is used here.
}
