% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_plots.R
\name{plot_decoded_data}
\alias{plot_decoded_data}
\title{Plot decoded data}
\usage{
plot_decoded_data(data, col, name = "none", title = "none")
}
\arguments{
\item{data}{Data vector to be plotted.}

\item{col}{Color of the states.}

\item{name}{optional, y-axis label, if "none" than "data".}

\item{title}{optional, title of the plot, if "none", then no title.}
}
\description{
Plot a data vector, colored with regard to the Viterbi decoded states.
}
