% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_params.R
\name{starize}
\alias{starize}
\title{Natural parameters to working parameters (theta->theta.star)}
\usage{
starize(theta, N, p, dists, scale_kappa = 1, zero_inf = FALSE)
}
\arguments{
\item{theta}{Vector of natural parameters.}

\item{N}{Number of states of the HMM.}

\item{p}{Vector of degree of autoregression within the distributions.}

\item{dists}{Vector of the distributions in the HMM.}

\item{scale_kappa}{Default 1, Scaling factor for kappa to avoid numerical issues in optimization for large kappa.}

\item{zero_inf}{Default FALSE, indicates if the gamma distributed variables should incorporate zero-inflation.}
}
\value{
Vector of working parameters of the AR(p)-HMM.
}
\description{
Transform the natural parameters of AR(p)-HMMs to their working parameters for optimization (theta -> theta.star).
Currently only works for distributions with exactly two parameters.
}
