% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_params.R
\name{unstarize}
\alias{unstarize}
\title{Working parameters to natural parameters (theta.star->theta)}
\usage{
unstarize(theta.star, N, p, dists, scale_kappa = 1, zero_inf = FALSE)
}
\arguments{
\item{theta.star}{Vector of working parameters.}

\item{N}{Number of states of the HMM.}

\item{p}{Vector of degree of autoregression within the distributions.}

\item{dists}{Vector of the distributions in the HMM.}

\item{scale_kappa}{Default 1, Scaling factor for kappa to avoid numerical issues in optimization for large kappa.}

\item{zero_inf}{Default FALSE, indicates if the gamma distributed variables should incorporate zero-inflation.}
}
\value{
List of natural parameters of the AR(p)-HMM.
}
\description{
Transform the working parameters of AR(p)-HMMs to their natural parameters for optimization (theta.star -> theta).
Currently only works for distributions with exactly two parameters.
}
